Struct fruently::retry_conf::RetryConf

pub struct RetryConf {
    // some fields omitted
}

You can calculate retrying interval as the following equation:

retry_interval = exp ** (multiplier + retry_counts)

see: https://github.com/jimmycuadra/retry/blob/v0.4.0/src/lib.rs#L142-L143

You can estimate to caluculate with concrete values like:

retry_counts: 10, e ^ (5 + 10)/1000.0/60.0/60.0 = 0.908060381242253, about 0.9 hour

retry_counts: 11, e ^ (5 + 11)/1000.0/60.0/60.0 = 2.4683640334744092, about 2.5 hours

retry_counts: 12, e ^ (5 + 12)/1000.0/60.0/60.0 = 6.709709098215361, about 6.7 hours

where multiplier = 5, e is exponential function.

Default values

multiplier: 5_f64

max: 10

Methods

impl RetryConf

fn new() -> RetryConf

fn max(self, max: u64) -> RetryConf

fn multiplier(self, multiplier: f64) -> RetryConf

fn build(self) -> (u64, f64)